--- routes/routes.js
@@ -1,101 +1,0 @@
-/**
- * Created by siroramirez on 23/05/17.
- */
-var users = require('./users');
-var express = require('express');
-var tokenMiddleware = require('../middleware/tokenVerification');
-
-// get an instance of the router for api routes
-var apiRoutes = express.Router();
-
-/**
- * @swagger
- * /register:
- *   post:
- *     summary: Registra un nou usuari
- *     tags: [User]
- *     consumes:
- *       - "application/json"
- *     produces:
- *       - "application/json"
- *     parameters: 
- *       - name: body
- *         in: body
- *         schema:
- *           $ref: "#/definitions/UserBody"
- *     responses:
- *       200:
- *         description: Operació executada amb éxit
- *         schema:
- *           $ref: "#/definitions/User"
- *       400:
- *         description: Ha hagut un error amb la operació
- *         schema:
- *           $ref: "#/definitions/Error"
- */
-apiRoutes.post('/register', users.createUser);
-
-apiRoutes.get('/', function (req, res) {
-  res.send("IntegrApp API Deployed!");
-});
-
-/**
- * @swagger
- * /users:
- *   get:
- *     summary: Retorna tots els usuaris registrats
- *     tags: [User]
- *     security:
- *       - user: []
- *     consumes:
- *       - "application/json"
- *     produces:
- *       - "application/json"
- *     responses:
- *       200:
- *         description: Operació executada amb éxit
- *         schema:
- *           type: array
- *           items:
- *             $ref: "#/definitions/User"
- *       400:
- *         description: Ha hagut un error amb la operació
- *         schema:
- *           $ref: "#/definitions/Error"
- *       403:
- *         description: Falta incloure el token
- *         schema:
- *           $ref: "#/definitions/Error"
- */
-apiRoutes.get('/users', tokenMiddleware.tokenCheck, users.getAllUsers);
-
-/**
- * @swagger
- * /login:
- *   post:
- *     summary: Login de l'usuari
- *     tags: [User]
- *     consumes:
- *       - "application/json"
- *     produces:
- *       - "application/json"
- *     parameters: 
- *       - name: body
- *         in: body
- *         schema:
- *           $ref: "#/definitions/LoginBody"
- *     responses:
- *       200:
- *         description: Operació executada amb éxit
- *         schema:
- *           $ref: "#/definitions/LoginResponse"
- *       400:
- *         description: Ha hagut un error amb la operació
- *         schema:
- *           $ref: "#/definitions/Error"
- */
-apiRoutes.post('/login', users.login);
-
-exports.assignRoutes = function (app) {
-  app.use('/api', apiRoutes);
-}
