+++ users/index.js
@@ -1,0 +1,7 @@
+// var routes = require('./routes');
+var tokenVerification = require('../middleware/tokenVerification');
+const routes = require('./routes');
+
+exports.assignRoutes = function (app) {
+  app.use('/api', routes.apiRoutes);
+}
